+++ CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Hovering charts highlights rank instead of points (#272)
 -   Better page titles (#275)
 -   Update Cypress version (#276)
+-   Remove number of parent tasks in TaskTable (#277)
 
 ## [0.6.0] - 2021-12-01